Whoa
I have added a lot of stuff. new palletes, new rule tiles, and have figured out how to get data from a tile map.
I use a dictionary and pull all the tiles on a specific tilemap and make them true, and use them for the specific item. Will look at making a scripatable asset out of this for each tag on the map. then assign it to map manager.